+++ src/main/java/basicstats/Numbers.java
@@ -5,11 +5,11 @@ import java.util.ArrayList;
 // Model for this assignment
 public class Numbers implements Model {
 
-  ArrayList<Integer> numbers;
+  ArrayList<Double> numbers;
   ArrayList<View> observers;
 
   public Numbers() {
-    numbers = new ArrayList<Integer>();
+    numbers = new ArrayList<Double>();
     observers = new ArrayList<View>();
   }
 
@@ -29,7 +29,7 @@ public class Numbers implements Model {
   }
 
   @Override
-  public void update(ArrayList<Integer> information) {
+  public void update(ArrayList<Double> information) {
     if (information.size() == 0) {
       numbers = information;
     } else {
@@ -39,11 +39,11 @@ public class Numbers implements Model {
     updateViews();
   }
 
-  public ArrayList<Integer> getData() {
+  public ArrayList<Double> getData() {
     return numbers;
   }
 
-  public void addData(Integer num) {
+  public void addData(Double num) {
     numbers.add(num);
 
     updateViews();
